+++ src/main/java/edu/psu/os/KV511/util/MessageUtil.java
@@ -17,7 +17,7 @@ public class MessageUtil {
 		in = new DataInputStream(socket.getInputStream());
 	}
 	
-	public void getRequest(int key) throws IOException {
+	public void getRequest(long key) throws IOException {
 		Message msg = new Message("GET", key, key);
 		System.out.println("GET Request "+ key);
 		out.writeUTF(msg.toString());
@@ -25,7 +25,7 @@ public class MessageUtil {
 		System.out.println("GET Response " + in.readUTF());
 	}
 	
-	public void putRequest(int key, int value) throws IOException {
+	public void putRequest(long key, long value) throws IOException {
 		Message msg = new Message("PUT", key, key);
 		System.out.println("Put Request "+ key);
 		out.writeUTF(msg.toString());
